srvFlask/requirements.txt
flask
pytesseract
numpy
opencv-python
pyspellchecker
deskew
qrcode